Initial upload of the WarDogs Alife Overhaul
WiP!! This is for WarDogs, my Anomaly full conversion. This does not include my entire alife system, but it may/not include enough to work with other modpacks. 

I do not give two shits about "sharing rights" 

Mods are for the people, by the people. Anyone who says you are not allowed to repost their work is simply seeking attention and affirmation. Feel FREE to share this where ever, with who ever. I do not care. Just mention me somewhere and i'll be happy.